Create a base class for live payloads
Create a new abstract class BaseLivePayload from LiveOSPayload.
